Cleanup
Moved common functionality from CommandHandler into Utils, phased out floats in favor of doubles, and improved reliability of command handling for game initialization and computer prompts.

+++ src/main/java/Model/Enterprise.java
@@ -18,8 +18,8 @@ public class Enterprise extends Entity {
 
     private byte lifeSupport;
 
-    private float warp = 5.0f;
-    private float energy = 5000.0f;
+    private double warp = 5.0f;
+    private double energy = 5000.0f;
 
     private int torpedoes;
