+++ src/main/java/Model/Sheild.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class Sheild {
     private String active;
     private int level;
-    private float units;
+    private double units;
 }
